Add files via upload
Figure 9 The comparison between our ST-MDA and state-of-the-art methods. The left three columns, from (a) to (c), are the labeled first frame (FirstFrm), discontinuous four test frames for (Test4Frm), and the ground-truth for Test4Frm (Truth), respectively. The rest columns from (d) to (i) are fire detection results. Among six fire detectors, both Rule and SP-UDA are unsupervised, while the rest ones are all supervised.