Add files via upload
Figure 9 The comparison between our ST-MDA and state-of-the-art methods on unknown fire-absent scenes. The left column (a), termed as NoFireFrm, is used for non-fire target scene. The second and third columns are test frames (Test4Frm) and their fire annotations (Truth). The rest columns from (d) to (i) are fire detection results. Among six fire detectors, both Rule and SP-UDA are unsupervised, while the others are all supervised.